incorporate redis as a persistence layer rather than a cache
this deleverages the entityMapper, preferring to read from cache instead of a
db. There will need to be some modification based on the valid hash key, which
will require some experimentation.

In addition, it will be important to develop a mitigation strategy for the
cache going down or stale for fault tolerance.