manage messages for an updated post
once the message processor processes the message, it should send back an
updated version of the post, which should then be written to the cache for the
next request.  This process may take a while, which is why it is managed async.

However, this addition breaks the universality of the service that was
introduced, so in a future version, perhaps this should be moved to its own
utility.