pass a timestamp with the generated object
in addition to a uid, the timestamp is also necessary for every posted object
in order for the db methods to function correctly. As it is written now, if
these are provided by the client, the values provided by the client are
preferred to the service generated values.

Additionally, there is an unstated dependence on lineage to be derivable from
these two attributes, so if a client wants to manage a different data structure
than a linear with time list, other attributes will need to be added to the
request to manage that structure (such as a parent and root).