provide both a id and a uid at request time
in order to ensure that there is a unique reference to the posted object, a uid
is created. It can still be overloaded by the client, but the id has to be
consistent across versions of an object to be processed, so uid gives a unique
reference to the posted object.