ensure the service generated values are written to the request object
when the request was unpacked after the server generated values were added,
there was a chance that the id, uid, and timestamp would be overwritten. In
this case, the id will be leveraged from the request or from the service if
it's not requested, but all other values will come from the service.

Another option on this might be to not unpack the request.